+++ tools/profileBuilder/model/transforms.go
@@ -23,7 +23,7 @@
 // that we will publish in this repository for general use. However, this tool
 // in the case that one has their own list of Services at given API Versions,
 // this may prove to be a useful tool for you.
-package main
+package model
 
 import (
 	"bytes"
@@ -41,8 +41,8 @@ import (
 	"os/exec"
 	"path"
 	"path/filepath"
-	"strings"
 	"sort"
+	"strings"
 	"time"
 
 	"github.com/marstr/collection"
@@ -81,7 +81,6 @@ const armPathModifier = "mgmt"
 // If not the empty string, this string should be stamped into files generated by the profileBuilder.
 // Note: This variable should be set by passing the argument "-X main.version=`{your value}`" to the Go linker. example: `go build -ldflags "-X main.version=f43d726b6e3f1e3eb7cbdba3982f0253000d5dc5"`
 var version string
-
 
 func main() {
 	var packages collection.Enumerator
@@ -221,6 +220,7 @@ func generateAliasPackages(packages collection.Enumerator) collection.Enumerator
 	packages = packages.Select(updateAliasPackageUserAgent)
 	return packages
 }
+
 //updateAliasPackageUserAgent updates the "UserAgent" function in the generated profile, if it is present.
 func updateAliasPackageUserAgent(x interface{}) interface{} {
 	cast, ok := x.(*alias)
@@ -262,9 +262,9 @@ func updateAliasPackageUserAgent(x interface{}) interface{} {
 	*retResults = updated
 
 	file := cast.ModelFile()
-	sort.SliceStable(file.Decls,  func(i, j int) bool {
+	sort.SliceStable(file.Decls, func(i, j int) bool {
 		return file.Decls[i].Pos() < file.Decls[j].Pos()
-	  })
+	})
 
 	return x
 }
@@ -363,8 +363,8 @@ func writeAliasPackages(x interface{}) interface{} {
 
 	var b bytes.Buffer
 	printer.Fprint(&b, files, file)
-	res, _ := imports.Process(outputPath,b.Bytes(), nil)
-	fmt.Fprintf(outputFile,"%s" , res)
+	res, _ := imports.Process(outputPath, b.Bytes(), nil)
+	fmt.Fprintf(outputFile, "%s", res)
 	return true
 }
